Statistically link libOTExtension, libMaliciousOTExtension, and libmiracl
Previously, the libraries libOTExtension and libMaliciousOTExtension were dynamically linked. But each of them was also statistically linked to libmiracl.
The resulting behaviour is not specified by the standard.
Luckily it worked on Linux.
But it does not work on macOS (and should not work on Windows either): https://stackoverflow.com/questions/31209693/static-library-linked-two-times
Concretely, the global variable mr_mip from the Miracl library was defined twice. Sometimes one was used, sometimes the other was used, depending from which part of the code (header / library) the Miracl functions were called.
This made `libscapi_sample` segfault in weird ways and finding the origin of the bug was extremely difficult.

This change might be required in other parts libscapi.

A better solution might to dynamically link all libraries.